# NodeJsNews/openapi.yaml
openapi: 3.0.0
info:
  title: API de Noticias
  version: 1.0.0
  description: API para gestionar noticias con búsqueda personalizada.
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
tags:
  - name: Noticias
    description: Operaciones relacionadas con la gestión de noticias
paths:
  /noticias:
    get:
      summary: Obtener todas las noticias
      tags:
        - Noticias
      responses:
        '200':
          description: Lista de noticias obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Noticia'
        '500':
          description: Error interno del servidor
    post:
      summary: Crear una nueva noticia
      tags:
        - Noticias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Noticia'
      responses:
        '201':
          description: Noticia creada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Noticia'
        '400':
          description: Error en la solicitud
  /noticias/filtro:
    get:
      summary: Buscar noticias con filtros personalizados
      tags:
        - Noticias
      parameters:
        - name: categoria
          in: query
          description: Filtrar por categoría
          required: false
          schema:
            type: string
        - name: idioma
          in: query
          description: Filtrar por idioma
          required: false
          schema:
            type: string
        - name: autor
          in: query
          description: Filtrar por autor
          required: false
          schema:
            type: string
        - name: fechaInicio
          in: query
          description: Fecha de inicio para el rango de búsqueda
          required: false
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          description: Fecha de fin para el rango de búsqueda
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Noticias encontradas con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Noticia'
        '404':
          description: No se encontraron noticias
        '500':
          description: Error interno del servidor
  /noticias/{id}:
    delete:
      summary: Eliminar una noticia por ID
      tags:
        - Noticias
      parameters:
        - name: id
          in: path
          description: ID de la noticia a eliminar
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Noticia eliminada con éxito
        '400':
          description: ID no válido
        '404':
          description: Noticia no encontrada
        '500':
          description: Error interno del servidor
components:
  schemas:
    Noticia:
      type: object
      properties:
        _id:
          type: string
        titulo:
          type: string
          description: Título de la noticia
        fuente:
          type: string
          description: Fuente de la noticia
        categoria:
          type: string
          description: Categoría de la noticia
        url:
          type: string
          description: URL de la noticia
        fecha:
          type: string
          format: date
          description: Fecha de publicación
        autor:
          type: string
          description: Autor de la noticia
        idioma:
          type: string
          description: Idioma de la noticia